Move options logging for errors into options.ts
